requirements.txt
certifi~=2023.11.17
fastapi~=0.109.0
Jinja2
pydantic~=2.5.3
pymongo~=4.6.1
uvicorn
websockets
numpy~=1.26.3
osmnx~=1.8.1
gunicorn
sse-starlette
asyncio~=3.4.3
networkx
pytz
python-dotenv
flask
PyJWT
requests
gunicorn
asgiref
